handle bad expressions
